refactor basic, edu
